// server/src/constants/env.type.ts
export enum EEnvType {
	NODE_ENV = 'NODE_ENV',
	PORT = 'PORT',
	GLOBAL_PREFIX = 'GLOBAL_PREFIX',
	SWAGGER_PATH = 'SWAGGER_PATH',
	SECRET_JWT = 'SECRET_JWT',
	REFRESH_TOKEN_SECRET_JWT = 'REFRESH_TOKEN_SECRET_JWT',
	DATABASE_HOST = 'DATABASE_HOST',
	DATABASE_USER = 'DATABASE_USER',
	DATABASE_PASSWORD = 'DATABASE_PASSWORD',
	DATABASE_PORT = 'DATABASE_PORT',
	DATABASE = 'DATABASE',
	HASH_ITERATIONS = 'HASH_ITERATIONS',
	HASH_SALT = 'HASH_SALT',
	PATH_FILE_UPLOAD = 'PATH_FILE_UPLOAD',
}